+++ EFT_S9/src/main/java/com/duoc/eft_s9/EFT_S9.java
@@ -1,16 +1,10 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- */
-
 package com.duoc.eft_s9;
 
-/**
- *
- * @author mvarg
- */
-public class EFT_S9 {
+import com.duoc.eft_s9.ui.MenuPrincipal;
 
+public class EFT_S9 {
     public static void main(String[] args) {
-        System.out.println("Hello World!");
+        MenuPrincipal menu = new MenuPrincipal();
+        menu.iniciar();
     }
 }